Move the dbgraph to its own new page.
It's the first page that's not in the blog. It's available on the top level
directory of the website as in the following URL:

   http://localhost:1313/compute-database-size/

Also, it's now using a Hugo shortcode rather than a partial. It's another
way to provide templates to building a page, targetting this time at page
written in markdown rather than other template work in HTML.